Reuse `KpiHome` view for embedded KPI details component
